requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.2
opencv-python>=4.8.0
scikit-learn>=1.3.0
Pillow>=10.0.0
numpy>=1.24.0
joblib>=1.3.0
plotly>=5.17.0
pandas>=2.0.0
